replace idptest with qtidptest
The original idptest isn't used any more.
