+++ idp/static/static_test.go
@@ -15,7 +15,7 @@ import (
 	"golang.org/x/net/context"
 
 	"github.com/CanonicalLtd/candid/idp"
-	idptest "github.com/CanonicalLtd/candid/idp/qtidptest"
+	"github.com/CanonicalLtd/candid/idp/idptest"
 	"github.com/CanonicalLtd/candid/idp/static"
 	candidtest "github.com/CanonicalLtd/candid/internal/qtcandidtest"
 	"github.com/CanonicalLtd/candid/store"
