+++ idp/usso/ussodischarge/ussodischarge_test.go
@@ -27,7 +27,7 @@ import (
 
 	"github.com/CanonicalLtd/candid/config"
 	"github.com/CanonicalLtd/candid/idp"
-	idptest "github.com/CanonicalLtd/candid/idp/qtidptest"
+	"github.com/CanonicalLtd/candid/idp/idptest"
 	"github.com/CanonicalLtd/candid/idp/usso/ussodischarge"
 	candidtest "github.com/CanonicalLtd/candid/internal/qtcandidtest"
 	"github.com/CanonicalLtd/candid/store"
